start implementing scene test

+++ src/rhi/rhi_view.cpp
@@ -1,6 +1,6 @@
 #include "rhi_view.hpp"
 #include "rhi.hpp"
 
-RhiView::RhiView(RHI_VIEW* handle, size_t resource_id) 
+RhiView::RhiView(RHI_VIEW* handle, int resource_id) 
 	: RhiImpl<RHI_VIEW>(handle)
 	, m_view_id(resource_id) {}
